Id datatype has been changed

+++ backend/src/main/java/dev/taskManager/backend/model/entity/Role.java
@@ -1,6 +1,7 @@
 package dev.taskManager.backend.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -26,7 +27,7 @@ public class Role implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @JsonProperty("role_name")
     @Column(name = "role_name", nullable = false, length = 20)
